Use postcss-cli for build
