+++ src/basscss.css
@@ -1,3 +1,4 @@
+/*! Basscss | http://basscss.com | MIT License */
 
 @import 'basscss-base-reset';
 @import 'basscss-base-forms';
@@ -23,5 +24,4 @@
 @import 'basscss-background-colors';
 
 @import 'basscss-defaults';
-
